Huge refactor of rendering elements on page
The initial motivation for this work is to fix some layout issues and
deviations from the Swiss QR-bill standard.

As this work took a rather circuitous route, I don't have a meaningful
story to tell in a sequence of self-contained commits, so I'm afraid
it's all lumped together in a single monolithic one. My apologies.

In order to understand the layout code I ended up refactoring it
completely. The essential part is replacing the old, monolithic, 454
line long, `QRBill::draw_bill` function, and similar related functions,
with a more structured approach. The main aspects of this are:

+ Replace all the magic numbers with named values clearly related to the
  standard and style guide.

+ Reuse the same code to render corresponding elements on the
  receipt/payment sides, rather than having separate versions without
  any connection between them.

+ Improve the signal to noise ratio of the code, by providing some
  abstractions for related and repeated tasks. I think that this can be
  improved further, but it will do for now.

+ Break the code into smaller functions which tell a clearer story about
  what is going on.

+ Use the type system to help deal with the different length units: mm,
  pt and svg UU. This can/should be improved significantly. In this
  part, the lack of float arithmetic in const functions was a big
  obstacle, because of my initial choice of storing the
  receipt/payment values in `const` structures.

Additionally, to help understand/evaluate the behaviour of the crate, I
have

+ added a local copy of the style guide currently available at https://www.six-group.com/dam/download/banking-services/standardization/qr-bill/style-guide-qr-bill-en.pdf

+ added 6 examples in `examples/from-style-guide.rs` corresponding to
  those found in the above guide

Things that are still obviously missing:

+ Support for alternative processes

+ Proper treatment of additional information (Ustrd and StrdBkginf)